feat: :recycle: Refactor docs new - Move FileUploadForm and UrlForm to own components
